// src/app/pokemon/interfaces/pokemon-data-for-table.interface.ts
// Generated by https://quicktype.io

export interface PokemonDataForTableInterface {
  id:                       number;
  name:                     string;
  base_experience:          number;
  height:                   number;
  order:                    number;
  weight:                   number;
  abilities:                Ability[];
  sprites:                  Sprites;
  onEdition:                boolean;
}

export interface Ability {
  is_hidden: boolean;
  slot:      number;
  ability:   NameUrl;
}

export interface NameUrl {
  name: string;
  url:  string;
}

export interface Sprites {
  other?:             Other;
}

export interface Other {
  dream_world:        DreamWorld;
}

export interface DreamWorld {
  front_default: string;
}